basejs/3.0/pushing/python/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
bidict==0.22.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
multidict==6.0.4
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.31.0
typing_extensions==4.6.2
urllib3==2.0.2
websocket-client==1.5.2
yarl==1.9.2
zipp==3.15.0
